agent-terminal-system/mcp-bridge/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pyjwt==2.8.0
redis==5.0.1
docker==6.1.3
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
mcp==0.1.0

# Optional but recommended
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2